Fix sign state issue for exponents
Can't reuse the Sign rule for exponents, so just do it normally. By
reusing the sign rule, it resets the sign integer used in parsing
numbers (the alternative is to use a stack here and push/pop signs but
it's not useful for this). When an exponent with a sign was parsed, it
consequently set the sign to 1 (positive).