Remove the peek method
This hasn't been used so far and I don't see it being used since
unread() exists. It might be necessary to add something to get any
unread rune back out of the lexer, but I don't see a scenario in which
it's useful to mix codf lexing and other reading.