Add duration tests
These are being added while working on duration parsing, as with
anything else. The success cases are written ahead of the actual
parsing code while failures are intended to hit specific parts of
coverage (i.e., to ensure the code is hitting the branches I expect),
since it's useful as a matter of seeing what code is run for a subtest.