Implement map parsing
Still limited to strings/barewords, but no more MapKey interface type
for the key.

This adds a MapEntry node type as well, specifically used to represent
the key-value pair in a Map. Each MapEntry retains the order of its
insertion (Ord), accounting for duplicates (i.e., if there are gaps in
a sequence of Ord, it's because there were duplicate keys).