Widen the bareword character range
Bareword initials are now permitted to be any non-separator,
non-decimal, and non-reserved character (and not a quote because
strings).

Bareword tails are about the same but allow decimals, pounds, hyphens,
and plus signs. This captures a pretty wide range of characters now.
I'm resisting the urge right now to allow escaping whitespace in
barewords just because that's embedding quote logic in an unquoted
context.