Replace ' comment with -- comment
This has some ambiguous behavior in the language since ' was previously
unused (and forbidden in barewords), whereas '--' is permitted in
a bareword. This means that foo--foo is not a comment but just
a bareword of "foo--foo", but if you put a space before the first
hyphen, it becomes a comment. So, clear separation is necessary with
this, where the apostrophe is pretty clear about its use.