More parser tests
Tried to flesh this out some more to test successful/failed parser
cases and made it a little easier to construct an AST to compare
against. This does skip testing exact token values, so it's not
comprehensive and focuses more on whether the general input/output is
correct. For example, token locations are tested somewhat in the lexer
and the parser doesn't care much about them except to present errors,
so it's just been omitted. The only real thing of interest is a token's
defined value, such as a boolean or string, rather than its metadata.