Remove some spacing requirements from parser
Whitespace before/after []s and #{}s is now optional. This simplifies
the grammar a bit. As a result of this, there are no longer any places
where whitespace is mandatory from the parser's point of view. It is
still a requirement if passing a map as the first element of
a statement, though: stmt#{}; is equivalent to stmt# { } ;.

Comparatively, stmt[]; is still unambiguous.