Add raw strings
This fills a kind of specific niche of wanting to possibly embed other
source code inside of a codf document (e.g., Lua, peculiar as that is).

To avoid the pain in the neck that is Go's un-quote-able ` inside of
raw strings, a double `` is a quoted `. I don't think this quoting
should be used in double-quoted strings, though -- they have their
escape form, and raw strings have theirs such that they minimize
possible escape-able runes.