Add Each and Select functions
These are used to work with Nodes to iterate over them and select
specific nodes. Each is indiscriminate walking and Select is for
grabbing children with specific names.

It occurs to me now that Each walks parameters but not the sub-nodes of
Map or Array right now, and I can't decide if this is desirable.
Something to think about.