Add walk.go
This adds the Walk function and Walker/WalkExiter interfaces to
recursively walk through an AST. This can make it fairly easy to
iterate over configs.

walk_test.go includes a sort of example (not yet an Example test). It's
mostly just for making sure Walk works as expected right now.