Add a ParseExpr method to Parser
This is useful during walks / iterations where you want to take
a particular expression in a statement, section, or elsewhere and
replace it with another parsed expression node.

For example, given the following statement:

    stmt %(FOO) %(BAR);

Say you expand both words to something else, like a quoted string or
a boolean or a number. You can then take the expanded text and hand it
to ParseExpr to get back an ExprNode to replace the original ExprNode
the expansion was from.

So, if FOO expands to `true` and BAR expands to `2#1100`, using
ParseExpr on each string will return an ExprNode that can then be used
to replace the original word expressions that were there. This leaves
you with the statement post-interpolation:

    stmt true 2#1100;

Which presumably means something in the context of the thing accepting
the statement.