Add a ParseExpr method to Parser
This is useful during walks / iterations where you want to take
a particular expression in a statement, section, or elsewhere and
replace it with another parsed expression node.

For example, given the following statement:

         stmt %(FOO) %(BAR);

Say you expand both words to something else, like a quoted string or
a boolean or a number. You can then take the expanded text and hand it
to ParseExpr to get back an ExprNode to replace the original ExprNode
the expansion was from.