Clean up some unnecessary code
The error under parseStatementSentinel mirrors the error in closeError,
but had an invalid format string, so there's no point having both when
one is wrong.

The error under (*mapBuilder).addExpr mirrored the prior one, but is
effectively impossible to encounter without directly modifying the
mapBuilder's key ExprNode, which isn't a concern since it's unexported.