Backport test cases
These are borrowed from the switch-comment-leader branch, so they
maintain the current comment form but lack the bareword bracing patches
(which will also be backported).

The tests are intended to increase coverage of a few specific parts of
the lexer and ensure that error cases in the parser are still errors.