Only use Map() node result if no error is returned
This fixes #12 by only using the Node result of a Map() call if the
Map() call was successful. Otherwise, if there's an error, it uses the
original node to describe the error.

In addition, it skips over nil children -- because a nil child cannot
be mapped, it simply ignores them instead of modifying their cells.