Fix child reassignment after Map
This fixes a bug that was introduced in e6ce13f9 where the newChild was
checked and added to the child list but was not walked after being
mapped. This resulted in invalid handling of the child, since the walk
would then proceed incorrectly with the original child instead of its
replacement.

This in particular breaks things like include mappings that replace
a statement with a document (or collection of documents).