Switch comment lexing to // instead of quote
This is more natural for most folks, though it means you can't have
'//' itself as a word. This isn't really an issue, but is a limitation.
You can, of course, just use "//" to create the same thing. Whether
it's the same thing to the thing walking a config is another matter.

In addition, whitespace tokens now preserve their raw values and
comments have both a raw value (this includes the `//`) and a string
value (no `//`). Comment tokens no longer consume the trailing newline,
meaning that `"//foo\n"` is two tokens: a comment and whitespace.

- Switches comment format from "'comment" to "//comment".

- Unreads the newline (or EOF) following a comment, ensuring that the
  newline following the comment is preserved as a token.

- Retains the run of whitespace for a TWhitespace token in the token's
  Raw field. Ideally, this means that printing the raw value of each
  token, one after the other, will reproduce the input document(s).