Implement lexer flags to disable special tokens
This makes some adjustments to the lexer to allow a user to disable
tokens they will never use. So, for example, durations might be
useless, so those can be disabled. All number can be disabled, which
implies durations, rationals, floats, and so on as well.

In order to support boolean lexing, those are no longer part of the
parser's conversion steps and have been moved into the lexer. So,
LexNoBools will produce words instead of boolean values when that's
useful (maybe you have a walk step that understands a very narrow set
of words to be booleans).

There isn't a LexNoInts right now, so you can't disable those without
disabling all numbers. This may be supported in the future.

- Add Flags field to Lexer to hold a LexerFlag value. These are simple
  bit flags -- any additional customization of their behavior should
  come through additional fields.

- Move wordToBool to the lexer so that the lexer is responsible for
  identifying booleans. It was a little weird that the parser was
  responsible for this. The less the parser needs to do to munge tokens
  into an acceptable state, the better.

- Add additional tests to check that the lexer will treat disabled
  tokens as words.

- Adds an additional test around lexer precision, since the lexer will
  now assign its defaults instead of leaving them at zero. Zero still
  gets you the default, however, which needs to be checked.

Closes #14.