Add support for ending statements on newline
This adds a bitflag to the parser to enable parsing newlines as
statement sentinels in place of semicolons. Semicolons continue to
work.

This still needs work to support a backslash before EOL as a means of
ending a statement, but since there's no lookahead on tokens right now,
that's a little tricky. Might be doable by checking if the last element
of the current statement is a '\' word.