Fix bug in comment->word lexing
When the lexer began parsing the start of a comment, after encountering
a '/' rune, it would then transition to a word token if the next rune
wasn't also a '/'. This always buffered a '/' rune, even when the
encountered rune was not a '/'. Presumably this is a side-effect of
moving to two-comments and buffering the leading "//" prefix of
comments in the lexer.