Add Filter, Last, and First functions
This uses a Filterer to select children of a ParentNode and return
them. For example, to get all foo statements in a document root, you'd
use `Filter(doc, FilterStatements, FilterNamed("foo"))`. This is
a convenience for grabbing relevant elements of an AST in an order
that's convenient for iterating thorugh, rather than walking the whole
document all at once.

Debatable if this belongs in codf or a package underneath it just for
filtering.