Add Filter, Last, and First functions under codx
codx is a new subpackage of codf that contains utility functions. Like
everything else about codf, the API is currently unstable but unlikely
to change drastically.

This uses a Filterer to select children of a ParentNode and return
them. For example, to get all foo statements in a document root, you'd
use `Filter(doc, FilterStatements, FilterNamed("foo"))`. This is
a convenience for grabbing relevant elements of an AST in an order
that's convenient for iterating thorugh, rather than walking the whole
document all at once.