Add AllOf and OneOf ExprValue implementations
These can be used to create first-matching or all-matching sets of
arguments. AllOf is often more useful for pulling both the location or
token of a node out at the same time as its value so that you can
produce an error message. OneOf is useful with **T values where it's
useful to use the first non-nil value.